requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.10
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.8.0
bs4==0.0.1
celery==5.2.7
certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
cssselect==1.1.0
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1
django-celery-beat==2.4.0
django-oauth-toolkit==2.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
idna==3.4
iniconfig==1.1.1
isort==5.10.1
jwcrypto==1.4.2
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.15.3
pyparsing==3.0.9
pytest==7.1.3
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
tzdata==2022.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
